Reduce MoE activation memory in DualPipeV
- Remove EpilogOuts/epilog_b (dead code; logits not needed after loss)
- Remove Stage2/Stage4 args and outs (only ctx needed for a2a backward)
- Add padded_index_gather: avoids saving input for backward + reduces
  CUDA allocator fragmentation via fixed-alignment output buffers
- Free stage-boundary tensor storage early via untyped_storage().resize_(0)
  after async all-to-all consumers complete (sorted_tokens after Stage 2,
  gathered_tokens after Stage 3, stage3 moe_outs after Stage 4)
- Add fwd_comm_deferred_free to ExecutionCtx for clean deferred-free in
  the overlapped path without changing stage function signatures
- Add layer_partition() for memory-aware pipeline stage assignment
- Add per-layer saved-tensor profiling with weight/activation distinction

+++ tests/test_fsdp.py
@@ -95,7 +95,7 @@ def apply_fsdp(model, mesh: torch.distributed.DeviceMesh, dtype):
     other_fsdp_mesh = mesh["dp", "ep"]._flatten()
     mp = MixedPrecisionPolicy(
         param_dtype=dtype,
-        reduce_dtype=dtype,
+        reduce_dtype=torch.float32,
         output_dtype=None,
         cast_forward_inputs=True,
     )
